network: move auth-related code to submodule
